% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_EP_output.R
\name{convert.timestamp}
\alias{convert.timestamp}
\title{Convert timestamp}
\usage{
convert.timestamp(df, year)
}
\arguments{
\item{df}{a data frame the simulation csv result eplusout.csv is read into using read.eplusout}

\item{year}{year of the simulation result. EnergyPlus output doesn't have year in the timestamp}
}
\value{
a data frame with `Date/Time` converted to POSIXct
}
\description{
Convert timestamp
}
