% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_EP_output.R
\name{read.eplusout}
\alias{read.eplusout}
\title{read EnergyPlus simulation eplusout.csv}
\usage{
read.eplusout(result.csv.dir, f)
}
\arguments{
\item{result.csv.dir}{folder containing the result csv, usually eplusout.csv}

\item{f}{filename, usually eplusout.csv}
}
\value{
a data frame containing emission.exfiltration, emission.exhaust,
    emission.ref, emission.rej, emission.surf, emission.overall corresponding
    to five AH component and the overall AH. Also containing energy.elec,
    energy.overall for energy consumption. The unit of AH and energy columns
    are J
}
\description{
read EnergyPlus simulation eplusout.csv
}
